pyproject.toml
[project]
name = "vibedeck"
version = "0.2.0"
description = "Live-updating transcript viewer, web frontend and tooling for Claude Code and OpenCode sessions"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [
    { name = "Tijs Zwinkels", email="tijs@tinkertank.eu" }
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sse-starlette>=2.0.0",
    "watchfiles>=0.21.0",
    "jinja2>=3.1.0",
    "markdown>=3.5.0",
    "nh3>=0.2.0",
    "click>=8.1.0",
    "click-default-group>=1.2.0",
    "pyyaml>=6.0.0",
    "tomli>=2.0.0;python_version<'3.11'",
    "ptyprocess>=0.7.0",
    "authlib>=1.3.0",
    "httpx>=0.27.0",
    "itsdangerous>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "black>=24.0.0",
]

[project.scripts]
vibedeck = "vibedeck:main"
claude-code-session-explorer = "vibedeck:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vibedeck"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.15.1",
]
